Saving work done so far.  But, abandoning this architecture.
I'm giving up on the IMGUI architecture.  It's too costly.

It can be made to work, but the cost is entirely too high.  I think
there are simpler architectures, such as that used by GEOS/64 for
instance.  GEOS/64 might represent a sweet-spot --- it's on the one hand
better factored and easier to get working because of a clear separation
of drawing and event reaction, while at the same time, retaining only
the bare minimum state necessary to get the interface working.

This is a pity, because I genuinely liked working with the IMGUI
interface from the programming standpoint.  But, I just couldn't get it
to work.  Visuals are terribly glitchy.  Commands are non-responsive.
Finally, the library implementation itself was starting to get rather
dense and hard to follow.  Chunks of code needed to be placed in exactly
the right order to get most of the visual glitches solved, for example.